backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
python-dotenv
httpx
pydantic
ultralytics
easyocr
transformers
torch
torchvision
